==> codes/03_analyse.Rmd <==
---
title: "03_Analyse des données"
author: "E.Maucuer & B.PATRA"
date: "12/24/2021"
output: html_document
fig_width: 10
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = F, message = F)
```

```{r libraries}
library(dplyr)
library(plotly)
library(ggplot2)
library(sp)
library(sf)
library(viridis)
library(xlsx)
source("02_plot_function.R")
```

```{r load data}
data_claims <- readRDS("./data/data_claims_year0.rds")
data_freq <- readRDS("./data/data_freq_year0.rds")
data_tot <- readRDS("./data/data_tot_year0.rds")
departement_shape <- st_read("./data/DEPARTEMENTS/DEPARTMENTS.shp")
population_dep <- read.xlsx("./data/ensemble.xlsx", sheetIndex = 2, startRow = 8, header = TRUE, colIndex = c(3,4,8,9))

```


```{r premieres etudes des data, echo = T}
summary(data_freq)
summary(data_claims)
```

# 1- Etude de la frequence 
```{r tab frequence}
prop.table(table(data_freq$claim_nb))
data_plot <- data_freq %>% group_by(claim_nb) %>% summarize(count = n())
data_plot$freq <- data_plot$count / sum(data_plot$count)
                                                            
plot_ly(data_plot) %>%
  add_trace(x= ~claim_nb, y= ~count, type = 'bar', color = ~as.factor(claim_nb), colors = "Reds") %>% 
  add_text(x= ~claim_nb, y= ~count, text= ~paste(round(freq*100,2),"%"), hoverinfo='none', textposition = 'top', showlegend = FALSE) 

```

## a- Barplots {.tabset}
### drv_age1
```{r plot fre drv_age1}
fun_plot(data_freq, "drv_age1", "claim_nb")
```
### drv_age1_G1
```{r plot fre drv_age1_G1}
fun_plot(data_freq, "drv_age1_G1", "claim_nb")
```
### drv_age1_G2
```{r plot fre drv_age1_G2}
fun_plot(data_freq, "drv_age1_G2", "claim_nb")
```
### drv_age1_G3
```{r plot fre drv_age1_G3}
fun_plot(data_freq, "drv_age1_G3", "claim_nb")
```

### drv_drv2
```{r plot fre drv_drv2}
fun_plot(data_freq, "drv_drv2", "claim_nb")
```
### drv_age2
```{r plot fre drv_age2}
fun_plot(data_freq, "drv_age2", "claim_nb")
```
### drv_age2_G1
```{r plot fre drv_age2_G1}
fun_plot(data_freq, "drv_age2_G1", "claim_nb")
```
### drv_age2_G2
```{r plot fre drv_age2_G2}
fun_plot(data_freq, "drv_age2_G2", "claim_nb")
```
### drv_age2_G3
```{r plot fre drv_age2_G3}
fun_plot(data_freq, "drv_age2_G3", "claim_nb")
```

### drv_age_lic1
```{r plot fre drv_age_lic1}
fun_plot(data_freq, "drv_age_lic1", "claim_nb")
```
### drv_age_lic1_G
```{r plot fre drv_age_lic1_G}
fun_plot(data_freq, "drv_age_lic1_G", "claim_nb")
```

### drv_age_lic2
```{r plot fre drv_age_lic2}
fun_plot(data_freq, "drv_age_lic2", "claim_nb")
```
### drv_age_lic2_G
```{r plot fre drv_age_lic2_G}
fun_plot(data_freq, "drv_age_lic2_G", "claim_nb")
```

### vh_age
```{r plot fre vh_age"}
fun_plot(data_freq, "vh_age", "claim_nb")
```
### vh_age_G1
```{r plot fre vh_age_G1}
fun_plot(data_freq, "vh_age_G1", "claim_nb")
```
### vh_age_G2
```{r plot fre vh_age_G2}
fun_plot(data_freq, "vh_age_G2", "claim_nb")
```
### vh_age_G3
```{r plot fre vh_ageG3}
fun_plot(data_freq, "vh_age_G3", "claim_nb")
```

### vh_cyl
```{r plot fre vh_cyl}
fun_plot(data_freq, "vh_cyl", "claim_nb")
```
### vh_din
```{r plot fre vh_din}
fun_plot(data_freq, "vh_din", "claim_nb")
```
### vh_fuel
```{r plot fre vh_fuel}
fun_plot(data_freq, "vh_fuel", "claim_nb")
```

### vh_make
```{r plot fre vh_make}
fun_plot(data_freq, "vh_make", "claim_nb")
```
### vh_make_G
```{r plot fre vh_make_G"}
fun_plot(data_freq, "vh_make_G", "claim_nb")
```

### vh_model
```{r plot fre vh_model}
fun_plot(data_freq, "vh_model", "claim_nb")
```
### vh_type
```{r plot fre vh_type}
fun_plot(data_freq, "vh_type", "claim_nb")
```

### vh_value
```{r plot fre vh_value}
fun_plot(data_freq, "vh_value", "claim_nb")
```
### vh_value_G1
```{r plot fre vh_value_G1}
fun_plot(data_freq, "vh_value_G1", "claim_nb")
```
### vh_value_G2
```{r plot fre vh_value_G2}
fun_plot(data_freq, "vh_value_G2", "claim_nb")
```
### vh_value_G3
```{r plot fre vh_value_G3}
fun_plot(data_freq, "vh_value_G3", "claim_nb")
```

### vh_weight
```{r plot fre vh_weight}
fun_plot(data_freq, "vh_weight", "claim_nb")
```

### risk_class_G
```{r plot fre risk_class_G}
fun_plot(data_freq, "risk_class_G", "claim_nb")
```

### pol_bonus
```{r plot fre pol_bonus}
fun_plot(data_freq, "pol_bonus", "claim_nb")
```
### bonus
```{r plot fre bonus}
fun_plot(data_freq, "bonus", "claim_nb")
```

### pol_payd
```{r plot fre pol_payd}
fun_plot(data_freq, "pol_payd", "claim_nb")
```

### pol_coverage
```{r plot fre pol_coverage}
fun_plot(data_freq, "pol_coverage", "claim_nb")
```

### pol_pay_freq
```{r plot fre pol_pay_freq}
fun_plot(data_freq, "pol_pay_freq", "claim_nb")
```
## b- Plots mean vs var
```{r plot mean var}
mean_variance(data_freq, 'drv_age1', 'claim_nb')
mean_variance(data_freq, 'drv_age_lic1', 'claim_nb')
mean_variance(data_freq, 'drv_age2', 'claim_nb')
mean_variance(data_freq, 'vh_age', 'claim_nb')
mean_variance(data_freq, 'vh_make', 'claim_nb')
mean_variance(data_freq, 'vh_din', 'claim_nb')
mean_variance(data_freq, 'vh_cyl', 'claim_nb')
```

# 2- Etude de la severite
```{r etude de claim amount}
summary(data_claims$claim_amount) 

boxplot(data_claims$claim_amount, ylim=c(0,2800), main = "Boxplot of claim amount")
boxplot(data_claims$Lclaim_amount, ylim=c(0,12), main = "Boxplot of log claim amount")


plot(ecdf(data_claims$claim_amount), xlim = c(0, 5000), main = "Fonction de répartition des montants de sinistre") 
plot(ecdf(data_claims$Lclaim_amount), xlim = c(0, 50), main = "Fonction de répartition du log des montants de sinistre") 

hist(data_claims$claim_amount, nclass = 10)
hist(data_claims$claim_amount, xlim = c(0,30000), nclass = 100)
hist(data_claims$claim_amount, xlim = c(0,10000), nclass = 1000, main = "Histogramme des montants de sinistres")
hist(data_claims$Lclaim_amount, main = "Histogramme du log des montants de sinistres")

names(which(table(data_claims$claim_amount) > 3))
```

```{r boxplots, fig_width = 20}
fun_boxplot(data_claims, "drv_age1_G1", "claim_amount", both = F)
fun_boxplot(data_claims, "drv_age1_G2", "claim_amount", both = F)

fun_boxplot(data_claims, "vh_age_G1", "claim_amount", both = F)
fun_boxplot(data_claims, "vh_age_G2", "claim_amount", both = F)
fun_boxplot(data_claims, "vh_value_G1", "claim_amount", both = F)
fun_boxplot(data_claims, "vh_value_G2", "claim_amount", both = F)
fun_boxplot(data_claims, "vh_value_G3", "claim_amount", both = F)
fun_boxplot(data_claims, "vh_make_G", "claim_amount", both = F)

fun_boxplot(data_claims, "pol_pay_freq", "claim_amount", both = F)
fun_boxplot(data_claims, "pol_coverage", "claim_amount", both = F)


```

```{r plot severite 3}
amountvsnb(data_claims$claim_amount)
amountvsnb_grouped(data_claims, "drv_age1_G1", "claim_amount")
amountvsnb_grouped(data_claims, "drv_age_lic1_G", "claim_amount")

amountvsnb_grouped(data_claims, "vh_fuel", "claim_amount")
amountvsnb_grouped(data_claims, "vh_value_G3", "claim_amount")
amountvsnb_grouped(data_claims, "vh_age_G2", "claim_amount")

amountvsnb_grouped(data_claims, "pol_pay_freq", "claim_amount")
amountvsnb_grouped(data_claims, "pol_coverage", "claim_amount")
```

# 3- Cartes
```{r}
data_by_dep <- data_tot %>%
  group_by(DEP) %>% 
  summarise(nb = sum(claim_nb), mean_claim = mean(claim_amount)) %>% 
  left_join(population_dep , by = c("DEP" = "Code.département")) %>% 
  mutate(nb_pour_10000 = (nb / Population.totale) * 100000)

data_plot <- data_by_dep %>% 
  left_join(departement_shape, 
          by = c("DEP" = "DEPT")) 


data_plot <- st_set_geometry(data_plot,data_plot$geometry)

plot(data_plot["nb_pour_10000"], pal = magma(10, direction = -1))

plot(data_plot["nb"], pal = magma(13, direction = -1))

claim_by_dep <- data_tot %>% 
  group_by(DEP) %>% 
  summarise(mean_claim = mean(claim_amount)) %>% 
  left_join(departement_shape, 
            by = c("DEP" = "DEPT"))

claim_by_dep <- st_set_geometry(claim_by_dep,claim_by_dep$geometry)

plot(claim_by_dep["mean_claim"],pal = magma(8, direction = -1))
```
